Revert "Run instant stack for every test"
This reverts commit 7080b9461b63e4fea41faf964aef79c38a01a316.
