+++ ui/package.json
@@ -33,7 +33,7 @@
     "chain:start:all": "concurrently -r -k \"yarn chain:sif\" \"yarn chain:eth\"",
     "core:build": "cd core && yarn build",
     "core:compile": "cd core && yarn compile",
-    "core:test:all": "yarn core:compile && yarn core:test",
+    "core:test:all": "yarn core:compile && concurrently -r -k -s first \"yarn stack:backend\" \"sleep 40 && yarn core:test\"",
     "core:test": "cd core && yarn test",
     "core:watch": "cd core && yarn watch",
     "lint-quick": "yarn pretty-quick --staged --pattern 'ui/**/*.{vue,ts,js,json}'",
@@ -59,13 +59,12 @@
     "@storybook/addon-essentials": "^6.2.0-beta.13",
     "@storybook/addon-links": "^6.2.0-beta.13",
     "@storybook/vue3": "^6.2.0-beta.13",
-    "@types/node": "^14.14.41",
     "concurrently": "^5.3.0",
     "ganache-cli": "^6.12.0",
     "husky": "^6.0.0",
     "lerna": "^3.22.1",
-    "pretty-quick": "^3.1.0",
     "serve": "^11.3.2",
+    "pretty-quick": "^3.1.0",
     "truffle": "^5.1.48",
     "vercel": "^21.0.1",
     "vue-loader": "^16.1.2",
