+++ ui/scripts/run-stack-backend.sh
@@ -5,5 +5,4 @@ killall sifnoded sifnodecli ebrelayer ganache-cli
 yarn concurrently -k -r -s first \
   "yarn chain:eth:revert" \
   "yarn wait-on tcp:localhost:7545 && yarn chain:sif:revert" \
-  "yarn wait-on http-get://localhost:1317/node_info tcp:localhost:7545 && yarn chain:peggy:revert" \
-  "yarn wait-on http-get://localhost:1317/node_info tcp:localhost:7545 && sleep 25 && echo \">> SIFCHAIN IS READY <<\" && sleep 86400" 
+  "yarn wait-on http-get://localhost:1317/node_info tcp:localhost:7545 && yarn chain:peggy:revert"
